+++ app/Filament/Admin/Resources/CreneauResource.php
@@ -96,7 +96,9 @@ class CreneauResource extends Resource
         return $table
             ->groups([
                 Group::make('date')->date()
-                ->collapsible(),
+                ->collapsible()
+                ->getDescriptionFromRecordUsing(fn (Creneau $record): string => Carbon::parse($record->date)->format('l')),
+
                 Group::make('creneau')
             ])
             ->defaultGroup('date')
@@ -157,28 +159,9 @@ class CreneauResource extends Resource
             ->actions([
                 Tables\Actions\Action::make('dissociate')
                     ->label('Libérer')
-                    ->action(fn($record) => self::dissociatePerm($record)),
-                Tables\Actions\Action::make('shotgun1')
-                    ->label(fn($record) => match($record->creneau) {
-                        'M' => '9h30-10h',
-                        'D' => '11h45-13h',
-                        'S' => '17h30-23h',
-                    })
-                        ->button()
-                        ->color(fn($record) => self::determineColor1($record))
-                        ->action(fn($record) => self::handleshotgun1($record)),
-                Tables\Actions\Action::make('shotgun2')
-                    ->label(fn($record) => match($record->creneau) {
-                        'M' => '10h-12h',
-                        'D' => '13h-14h30',
-                        'S' => '18h30-23h',
-                    })
                     ->button()
-                    ->color(fn($record) => self::determineColor2($record))
-                    ->action(fn($record) => self::handleshotgun2($record)),
-                    //->disabled(true), -> à creuser pour etre encore mieux
-                ])
-
+                    ->action(fn($record) => self::dissociatePerm($record)),
+            ])
             ->bulkActions([
                 Tables\Actions\BulkActionGroup::make([
                 ]),
@@ -237,183 +220,6 @@ class CreneauResource extends Resource
         Creneau::where('id', '=', $record->id)->update(['perm_id' => null]);
     }
 
-    /**
-     * Handle the "shotgun 1" action for a slot.
-     *
-     * @param mixed $record
-     * @return void
-     */
-    private static function handleshotgun1($record)
-    {
-        $astreinteType = null;
-        if ($record->creneau == "M") {
-            $astreinteType = "Matin 1";
-        } elseif ($record->creneau == "D") {
-            $astreinteType = "Déjeuner 1";
-        } elseif ($record->creneau == "S") {
-            $astreinteType = "Soir 1";
-        }
-        if ($astreinteType) {
-            $astreinteUser = Astreinte::where('creneau_id', $record->id)
-                ->where('member_id', 1) //A changer Filament::auth()->id()
-                ->where('astreinte_type', $astreinteType)
-                ->first();
-            if ($astreinteType=="Soir 1"){
-                $existingAstreinte = Astreinte::where('creneau_id', $record->id)
-                    ->first() !=null;
-            }
-            else {
-                $existingAstreinte = Astreinte::where('creneau_id', $record->id)
-                        ->where('astreinte_type', $astreinteType)->first() != null;
-            }
-            if (!$existingAstreinte) {
-                $astreinte = new Astreinte([
-                    'member_id' => 1, // À CHANGER
-                    'creneau_id' => $record->id,
-                    'astreinte_type' => $astreinteType,
-                ]);
-
-                // Enregistre l'instance dans la base de données
-                $astreinte->save();
-            }
-            else {
-                if ($astreinteUser) Astreinte::where('creneau_id', $record->id)
-                    ->where('member_id', 1) //A changer
-                    ->where('astreinte_type', $astreinteType)
-                    ->first()
-                    ->delete();
-               else Notification::make()
-                    ->title('Il n\'y a plus de places pour cette astreinte')
-                    ->color('danger')
-                    ->send();
-            }
-        }
-    }
-
-    /**
-     * Handle the "shotgun 2" action for a slot.
-     *
-     * @param mixed $record
-     * @return void
-     */
-
-    private static function handleshotgun2($record)
-    {
-        $astreinteType=null;
-        $astreinteUserOther=null;
-        if ($record->creneau=="M") {
-            $astreinteType = "Matin 2";
-        }
-        elseif ($record->creneau=="D") {
-            $astreinteType = "Déjeuner 2";
-        }
-        elseif ($record->creneau=="S") {
-            $astreinteType = "Soir 2";
-        }
-        if ($astreinteType) {
-            if ($astreinteType== "Soir 2"){
-                $existingAstreinte = Astreinte::where('creneau_id', $record->id)->count()>=3;
-                $astreinteUser = Astreinte::where('creneau_id', $record->id)
-                    ->where('member_id', 1) //A changer Filament::auth()->id()
-                    ->where('astreinte_type', $astreinteType)
-                    ->first();
-                $astreinteUserOther = Astreinte::where('creneau_id', $record->id)
-                    ->where('member_id', 1) //A changer Filament::auth()->id()
-                    ->first();
-            }
-            else {
-                $existingAstreinte = Astreinte::where('creneau_id', $record->id)
-                        ->where('astreinte_type', $astreinteType)->first() != null;
-                $astreinteUser = Astreinte::where('creneau_id', $record->id)
-                    ->where('member_id', 1)
-                    ->where('astreinte_type', $astreinteType)//A changer Filament::auth()->id()
-                    ->first();
-            }
-            if (!$existingAstreinte && !$astreinteUser && !$astreinteUserOther) {
-                $astreinte = new Astreinte([
-                    'member_id' => 1, // À CHANGER Filament::auth()->id()
-                    'creneau_id' => $record->id,
-                    'astreinte_type' => $astreinteType,
-                ]);
-                // Enregistre l'instance dans la base de données
-                $astreinte->save();
-            }
-            else{
-                if ($astreinteUser){
-                    Astreinte::where('creneau_id', $record->id)
-                        ->where('member_id', 1) //A changer Filament::auth()->id()
-                        ->where('astreinte_type', $astreinteType)
-                        ->first()
-                        ->delete();
-                }
-                elseif ($astreinteUserOther){
-                    Notification::make()
-                        ->title('Vous avez déjà une perm du soir')
-                        ->color('danger')
-                        ->send();
-                }
-                elseif ($existingAstreinte) {
-                Notification::make()
-                    ->title('Il n\'y a plus de places pour cette astreinte')
-                    ->color('danger')
-                    ->send();
-            }
-            }
-            }
-        }
-
-    /**
-     * Determine the color for "shotgun 1" action.
-     *
-     * @param mixed $record
-     * @return string|null
-     */
-    private static function determineColor1($record)
-    {
-        if ($record->creneau == "M") {
-            $astreinteType = "Matin 1";
-        } elseif ($record->creneau == "D") {
-            $astreinteType = "Déjeuner 1";
-        } elseif ($record->creneau == "S") {
-            $astreinteType = "Soir 1";
-        }
-        if (Astreinte::where('creneau_id', $record->id)
-            ->where('member_id', 1) //A changer Filament::auth()->id()
-            ->where('astreinte_type', $astreinteType)
-            ->first())
-                return 'success';
-        if (Astreinte::where('creneau_id', $record->id)
-            ->where('astreinte_type', $astreinteType)->first()) {
-            return 'danger';
-        }
-
-    }
-
-    /**
-     * Determine the color for "shotgun 2" action.
-     *
-     * @param mixed $record
-     * @return string|null
-     */
-    private static function determineColor2($record)
-    {
-        if ($record->creneau == "M") {
-            $astreinteType = "Matin 2";
-        } elseif ($record->creneau == "D") {
-            $astreinteType = "Déjeuner 2";
-        } elseif ($record->creneau == "S") {
-            $astreinteType = "Soir 2";
-        }
-        if (Astreinte::where('creneau_id', $record->id)
-            ->where('member_id', 1)
-            ->where('astreinte_type', $astreinteType) //A changer Filament::auth()->id()
-            ->first()) return 'success';
-        if (($astreinteType=="Soir 2" && Astreinte::where('creneau_id', $record->id)->count()>=3)||($astreinteType!="Soir 2" && Astreinte::where('creneau_id', $record->id)
-            ->where('astreinte_type', $astreinteType)->first())){
-            return 'danger';
-        }
-
-    }
 }
